fix: PUT 요청이 2번 왔을때 예외처리 (#155)
* fix: 이미 2fa등록 절차가 완료되었을 때 409 conflict를 반환하도록 수정

+++ backend/app/src/auth/two-factor.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common'
+import { Injectable, ConflictException } from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
 import { Repository } from 'typeorm'
 import { User } from 'user/user.entity'
@@ -15,9 +15,14 @@ export class TwoFactorService {
   ) {}
 
   async enable(uid: number): Promise<string> {
+    if (
+      (await this.twoFactorRepository.findOneBy({ user: { uid } })) !== null
+    ) {
+      throw new ConflictException('Two factor is already enabled')
+    }
+
     const user = await this.userRepository.findOneBy({ uid })
     user.twoFactor = true
-
     const twoFactor = new TwoFactor()
     twoFactor.user = user
     twoFactor.secret = authenticator.generateSecret()
@@ -33,8 +38,13 @@ export class TwoFactorService {
   }
 
   async disable(uid: number): Promise<void> {
+    if ((await this.twoFactorRepository.findBy({ user: { uid } })) === null) {
+      throw new ConflictException('Two factor is already disabled')
+    }
+
     const user = await this.userRepository.findOneBy({ uid })
     user.twoFactor = false
+
     await this.twoFactorRepository.delete({ user: { uid } })
     await this.userRepository.save(user)
   }
